Add escape hatch for large responses
