+++ src/app.js
@@ -74,77 +74,6 @@ exports.lambdaHandler = async function lambdaHandler(event, context, callback) {
     );
   }
 
-  if (
-    request.headers.accept &&
-    request.headers.accept.some(header => header.value.includes("text/html"))
-  ) {
-    const response = await renderPretty(pathname, result);
-    return callback(null, response);
-  } else {
-    // If we're not displaying HTML, proxy the body instead of redirecting.
-    const fetch = require("node-fetch");
-    const res = await fetch(url);
-    const lambdaRes = await fetch2LambdaResponse(res);
-    return callback(null, lambdaRes);
-  }
+  const response_ = await renderPretty(request, pathname, result);
+  return callback(null, response_);
 };
-
-const blacklistedHeaders = [
-  // https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/lambda-requirements-limits.html#lambda-blacklisted-headers
-  "Connection",
-  "Expect",
-  "Keep-alive",
-  "Proxy-Authenticate",
-  "Proxy-Authorization",
-  "Proxy-Connection",
-  "Trailer",
-  "Upgrade",
-  "X-Accel-Buffering",
-  "X-Accel-Charset",
-  "X-Accel-Limit-Rate",
-  "X-Accel-Redirect",
-  "X-Amz-Cf-*",
-  "X-Amzn-*",
-  "X-Cache",
-  "X-Edge-*",
-  "X-Forwarded-Proto",
-  "X-Real-IP",
-  // Read only headers
-  "Content-Length",
-  "Host",
-  "Transfer-Encoding",
-  "Via",
-  // Read-only Headers for CloudFront Origin Request Events
-  "Accept-Encoding",
-  "If-Modified-Since",
-  "If-None-Match",
-  "If-Range",
-  "If-Unmodified-Since",
-  "Range",
-  // Others
-  "x-github-request-id",
-  "x-fastly-request-id",
-  "x-geo-block-list",
-  "x-served-by"
-].map(n => n.toLowerCase());
-
-function isBlacklistedHeader(name) {
-  return blacklistedHeaders.indexOf(name.toLowerCase()) >= 0;
-}
-
-async function fetch2LambdaResponse(res) {
-  if (res.status === 404) return response.notFound();
-  // TODO(ry) use res.arrayBuffer() instead of res.text().
-  const body = await res.text();
-  const headers = {};
-  res.headers.forEach(function(value, name) {
-    if (!isBlacklistedHeader(name)) {
-      headers[name] = [{ key: name, value }];
-    }
-  });
-  return {
-    status: res.status,
-    body,
-    headers
-  };
-}
